+++ Sources/SwifterSwift/UIKit/UICollectionViewExtensions.swift
@@ -147,6 +147,33 @@ public extension UICollectionView {
         register(UINib(nibName: identifier, bundle: bundle), forCellWithReuseIdentifier: identifier)
     }
 
+    /// SwifterSwift: Safely scroll to possibly invalid IndexPath
+    ///
+    /// - Parameters:
+    ///   - indexPath: Target IndexPath to scroll to
+    ///   - scrollPosition: Scroll position
+    ///   - animated: Whether to animate or not
+    func safeScrollToItem(at indexPath: IndexPath, at scrollPosition: UICollectionView.ScrollPosition, animated: Bool) {
+        guard indexPath.item >= 0 &&
+            indexPath.section >= 0 &&
+            indexPath.section < numberOfSections &&
+            indexPath.item < numberOfItems(inSection: indexPath.section) else {
+                return
+        }
+        scrollToItem(at: indexPath, at: scrollPosition, animated: animated)
+    }
+
+    /// SwifterSwift: Check whether IndexPath is valid within the CollectionView
+    ///
+    /// - Parameter indexPath: An IndexPath to check
+    /// - Returns: Boolean value for valid or invalid IndexPath
+    func isValidIndexPath(_ indexPath: IndexPath) -> Bool {
+        return indexPath.section >= 0 &&
+            indexPath.item >= 0 &&
+            indexPath.section < numberOfSections &&
+            indexPath.item < numberOfItems(inSection: indexPath.section)
+    }
+
 }
 
 #endif