+++ Tests/SwiftStdlibTests/XCTestManifests.swift
@@ -52,11 +52,6 @@ extension CharacterExtensionsTests {
     static let __allTests__CharacterExtensionsTests = [
         ("testInt", testInt),
         ("testIsEmoji", testIsEmoji),
-        ("testIsLetter", testIsLetter),
-        ("testIsLowercased", testIsLowercased),
-        ("testIsNumber", testIsNumber),
-        ("testIsUpercased", testIsUpercased),
-        ("testIsWhiteSpace", testIsWhiteSpace),
         ("testLowercased", testLowercased),
         ("testOperators", testOperators),
         ("testRandom", testRandom),
@@ -233,6 +228,7 @@ extension SequenceExtensionsTests {
         ("testRejectWhere", testRejectWhere),
         ("testSingle", testSingle),
         ("testSum", testSum),
+        ("testWithoutDuplicates", testWithoutDuplicates),
     ]
 }
 
@@ -303,6 +299,8 @@ extension StringExtensionsTests {
         ("testIsAlphaNumeric", testIsAlphaNumeric),
         ("testIsDigits", testIsDigits),
         ("testIsNumeric", testIsNumeric),
+        ("testIsPalindrome", testIsPalindrome),
+        ("testIsSpelledCorrectly", testIsSpelledCorrectly),
         ("testisValidEmail", testisValidEmail),
         ("testIsValidFileURL", testIsValidFileURL),
         ("testIsValidHttpsUrl", testIsValidHttpsUrl),
@@ -310,6 +308,7 @@ extension StringExtensionsTests {
         ("testIsValidSchemedUrl", testIsValidSchemedUrl),
         ("testIsValidUrl", testIsValidUrl),
         ("testIsWhiteSpaces", testIsWhiteSpaces),
+        ("testItalic", testItalic),
         ("testLastCharacter", testLastCharacter),
         ("testLastDeletingLastPathComponent", testLastDeletingLastPathComponent),
         ("testLastDeletingPathExtension", testLastDeletingPathExtension),